Move chest tests to a separate class
